planning phases introduced

+++ FijiProjectInventory/ServiceLayer/PurchaseItemsServices.cs
@@ -4,22 +4,26 @@ using System.Linq;
 using FijiProjectInventory.Models;
 using FijiProjectInventory.Utilities;
 using System.Diagnostics;
+using System.Web;
+using Microsoft.AspNet.Identity;
+using FijiProjectInventory.Exceptions;
 
 namespace FijiProjectInventory.ServiceLayer
 {
     public class PurchaseItemsServices
     {
-        public static IEnumerable<RequiredItem> GetRequiredItems(int categoryId, DateTime projectDate)
+        public static IEnumerable<RequiredItem> GetRequiredItems(int categoryId, int projectDateId)
         {
             IList<RequiredItem> returnVar;
-            IEnumerable<IGrouping<int, UseByYearVw>> useByYear;
-            IList<int> pastYears;
-            using (InventoryEntitiesConnect db = new InventoryEntitiesConnect())
+            IEnumerable<IGrouping<int, UsedByProjectDate>> useByProjectDate;
+            IList<int> pastProjectIds;
+            var today = DateTime.Today;
+            using (InventoryEntities db = new InventoryEntities())
             {
                 returnVar =
-                (from p in db.sp_purchased_items_indate(projectDate)
+                (from p in db.sp_purchased_items_indate(projectDateId)
                  where p.CategoryId == categoryId
-                 orderby p.PurchaseId
+                 orderby p.SubcatId
                  select new RequiredItem
                  {
                      Boxes = p.Boxes,
@@ -29,28 +33,42 @@ namespace FijiProjectInventory.ServiceLayer
                      ItemRequiresRefridge = p.RequiresRefrig,
                      ItemsPerBox = p.ItemsPerBox,
                      PricePerBox = p.PricePerBox,
-                     PurchaseId = p.PurchaseId,
-                     SubcategoryDescription = p.Description,
-                     SubcategoryId = p.SubcatId,
+                     PurchaseId = p.PurchaseId, 
+                     ItemSubcategoryDescription = p.Description,
+                     ItemSubcategoryId = p.SubcatId,
                      SupplierId = p.SupplierId,
                      SupplierName = p.SupplierName,
                      RemainingItems = p.ItemsRemaining ?? 0
                  }).ToList();
-                IEnumerable<int> ids = returnVar.Select(r => r.PurchaseId);
-                useByYear = (from u in db.UseByYearVws
-                             where ids.Contains(u.PurchaseId)
-                             orderby u.PurchaseId
-                             group u by u.PurchaseId into g
-                             select g).ToList();
-                pastYears = GetPastYears(db);
-            }
-            if (useByYear.Any())
-            {
-                Debug.Assert(returnVar.Select(r => r.PurchaseId).MembersAreEqual(useByYear.Select(u => u.Key)), "Assertion that used by year and purchases have same id in same order FAILED");
-                int i = 0;
-                foreach (var u in useByYear)
+                useByProjectDate = (from u in db.UsedByProjectDates
+                                     group u by u.ItemSubcategoryId into g
+                                     select g).ToList();
+                pastProjectIds = (from p in db.ProjectDates
+                                  where p.StartDate < today
+                                 orderby p.StartDate descending
+                                 select p.Id).ToList();
+
+                byte[] sessionId = GetSessionId();
+                var ur = db.UserReviews.Find(sessionId);
+                if (ur == null)
                 {
-                    returnVar[i].UsedInYear = MapYears(pastYears, u);
+                    ur = new UserReview
+                    {
+                        SessionId = sessionId,
+                        UserId = GetUserId(),
+                        FirstViewTime = DateTime.UtcNow
+                    };
+                    db.UserReviews.Add(ur);
+                    db.SaveChangesAsync();
+                }
+            }
+            if (useByProjectDate.Any())
+            {
+                Debug.Assert(returnVar.Select(r => r.ItemSubcategoryId).MembersAreEqual(useByProjectDate.Select(u => u.Key)), "Assertion that used by year and purchases have same id in same order FAILED");
+                int i = 0;
+                foreach (var u in useByProjectDate)
+                {
+                    returnVar[i].UsedInYear = MapYears(pastProjectIds, u);
                     ++i;
                 }
             }
@@ -59,30 +77,73 @@ namespace FijiProjectInventory.ServiceLayer
         public static RequiredItemsModel GetRequiredItemsModel()
         {
             var returnVar = new RequiredItemsModel();
-            using (InventoryEntitiesConnect db = new InventoryEntitiesConnect())
+            using (InventoryEntities db = new InventoryEntities())
             {
-                returnVar.PastYears = GetPastYears(db);
+                returnVar.PastProjectDates = (from pd in db.ProjectDates
+                                              orderby pd.StartDate descending
+                                              select pd.StartDate).ToList();
                 returnVar.Suppliers = db.Suppliers.ToKVEnumerable(s => s.Id, s => s.Name);
                 returnVar.Categories = db.Categories.ToKVEnumerable(c=>c.Id, c=>c.Name);
-                returnVar.ProjectDates = db.ProjectDates.Select(pd=>pd.StartDate).ToList();
+                returnVar.ProjectDates = db.ProjectDates.ToKVEnumerable(pd=>pd.Id,pd=>pd.StartDate);
             }
             return returnVar;
         }
-        static public void AddOrUpdateItems (IEnumerable<RequiredItem> data, byte categoryId)
+        static byte[] GetSessionId()
         {
-            using (InventoryEntitiesConnect db = new InventoryEntitiesConnect())
+            return Convert.FromBase64String(HttpContext.Current.Session.SessionID);
+        }
+        static Guid GetUserId()
+        {
+            return Guid.Parse(HttpContext.Current.User.Identity.GetUserId());
+        }
+        static public void AddOrUpdateItems (IEnumerable<RequiredItem> data, byte categoryId, byte projectDateId)
+        {
+            Debug.Assert(data != null && data.Any(), "Assertion that add or update item contains at least 1 entry failed");
+            using (InventoryEntities db = new InventoryEntities())
             {
+                var rv = db.UserReviews.AsNoTracking().FirstOrDefault(ur=>ur.SessionId == GetSessionId()); //no tracking so it does not update on every call to save
+                if (rv == null)
+                {
+                    Elmah.ErrorSignal.FromCurrentContext().Raise(new UnexpectedSessionIdChange("beginning CRUD operation, sessionID has changed from last data view"));
+                    rv = (from ur in db.UserReviews
+                          where ur.UserId == GetUserId()
+                          orderby ur.FirstViewTime descending
+                          select ur).First();
+                }
+                
                 foreach (var d in data)
                 {
                     if (d._destroy)
                     {
-                        if (d.PurchaseId != 0)
+                        bool toDelete = false;
+                        if (d.PurchaseId != 0 && 
+                            !db.Purchases.Any(p => p.StoreMovements.Any() && p.ProjectDate == null))
                         {
                             var p = new Purchase { Id = d.PurchaseId };
                             db.Purchases.Attach(p);
                             db.Purchases.Remove(p);
                             db.SaveChanges();
+                            toDelete = true;
                         }
+                        if (d.ItemSubcategoryId != 0 && 
+                            !db.Purchases.Any(p => p.ItemSubcategoryId==d.ItemSubcategoryId))//p.Id != d.PurchaseId && 
+                        {
+                            var s = new ItemSubcategory { Id = d.ItemSubcategoryId };
+                            db.ItemSubcategories.Attach(s);
+                            db.ItemSubcategories.Remove(s);
+                            db.SaveChanges();
+                            toDelete = true;
+                        }
+                        if (d.ItemId != 0 && 
+                            !db.ItemSubcategories.Any(s =>s.ItemId == d.ItemId)) // s.Id!=d.SubcategoryId && 
+                        {
+                            var i = new Item { Id = d.ItemId };
+                            db.Items.Attach(i);
+                            db.Items.Remove(i);
+                            db.SaveChanges();
+                            toDelete = true;
+                        }
+                        if (toDelete) { rv.Deletes++; }
                     } else {
                         Item i;
                         if (d.ItemId == 0)
@@ -103,17 +164,17 @@ namespace FijiProjectInventory.ServiceLayer
                         i.Notes = d.ItemNotes;
                         i.RequiresRefrig = d.ItemRequiresRefridge;
 
-                        Subcategory s;
-                        if (d.SubcategoryId == 0)
+                        ItemSubcategory s;
+                        if (d.ItemSubcategoryId == 0)
                         {
-                            s = new Subcategory();
-                            db.Subcategories.Add(s);
+                            s = new ItemSubcategory();
+                            db.ItemSubcategories.Add(s);
                         }
                         else
                         {
-                            s = db.Subcategories.Find(d.SubcategoryId);
+                            s = db.ItemSubcategories.Find(d.ItemSubcategoryId);
                         }
-                        s.Description = d.SubcategoryDescription ?? string.Empty;
+                        s.Description = d.ItemSubcategoryDescription ?? string.Empty;
                         s.ItemId = d.ItemId;
                         s.Item = i;
 
@@ -122,52 +183,43 @@ namespace FijiProjectInventory.ServiceLayer
                         {
                             p = new Purchase();
                             db.Purchases.Add(p);
+                            rv.Creates++;
                         }
                         else
                         {
                             p = db.Purchases.Find(d.PurchaseId);
+                            rv.Updates++;
                         }
                         p.ItemsPerBox = (byte)d.ItemsPerBox;
                         p.PricePerBox = d.PricePerBox;
-                        p.SubcategoryId = d.SubcategoryId;
-                        p.Subcategory = s;
+                        p.ItemSubcategoryId = d.ItemSubcategoryId;
+                        p.ItemSubcategory = s;
                         p.SupplierId = d.SupplierId;
+                        p.PurchaseForProjectDateId = projectDateId;
                         p.Boxes = d.Boxes;
 
                         db.SaveChanges(); //todo savechangesasync
 
                         d.ItemId = i.Id;
-                        d.SubcategoryId = s.Id;
+                        d.ItemSubcategoryId = s.Id;
                         d.PurchaseId = p.Id;
                     }
                 }
+                db.UserReviews.Attach(rv);
+                db.SaveChangesAsync();
             }
         }
-        public IEnumerable<int> GetPastYears()
+
+        private static IList<int> MapYears(IList<int> ids, IEnumerable<UsedByProjectDate> usedByProject)
         {
-            using (InventoryEntitiesConnect db = new InventoryEntitiesConnect())
-            {
-                return GetPastYears(db);
-            }
-        }
-        private static List<int> GetPastYears(InventoryEntitiesConnect db)
-        {
-            return (from p in db.Purchases
-                    where p.DatePurchased.HasValue
-                    group p by p.DatePurchased.Value.Year into yrs
-                    orderby yrs.Key descending
-                    select yrs.Key).ToList();
-        }
-        private static IList<int> MapYears(IList<int> years, IEnumerable<UseByYearVw> usedByYear)
-        {
-            Debug.Assert(years.MembersUniqueDescending());
-            Debug.Assert(usedByYear.Select(u => u.YearUsed).MembersUniqueDescending(), "Assertion that years for a given purchase are descending failed");
-            var returnVar = new List<int>(years.Count);
+            Debug.Assert(ids.MembersUniqueDescending(), "Assertion that ids of past project dates are unique and descending failed");
+            Debug.Assert(usedByProject.Select(u => u.StartDate).MembersUniqueDescending(), "Assertion that years for a given purchase are descending failed");
+            var returnVar = new List<int>(ids.Count);
             int i = 0;
-            foreach (var u in usedByYear)
+            foreach (var u in usedByProject)
             {
-                while (years[i] != u.YearUsed) { returnVar.Add(0); }
-                returnVar.Add(years[i++]);
+                while (ids[i] != u.ItemSubcategoryId) { returnVar.Add(0); }
+                returnVar.Add(ids[i++]);
             }
             return returnVar;
         }
